refactor: Organize series DTO package structure
