+++ data/src/main/java/com/madrid/data/dataSource/remote/mapper/SeriesMappers.kt
@@ -5,18 +5,17 @@ import com.madrid.data.dataSource.remote.dto.series.SearchSeriesResponse
 import com.madrid.data.dataSource.remote.dto.series.SeasonsNetwork
 import com.madrid.data.dataSource.remote.dto.series.SeriesCastNetwork
 import com.madrid.data.dataSource.remote.dto.series.SeriesDetailsResponse
-import com.madrid.data.dataSource.remote.dto.series.SeriesGenres
 import com.madrid.data.dataSource.remote.dto.series.SeriesResult
 import com.madrid.data.dataSource.remote.dto.series.SeriesReviewResult
 import com.madrid.data.dataSource.remote.dto.series.SimilarSeriesNetwork
-import com.madrid.data.dataSource.remote.response.series.AiringTodaySeriesResult
-import com.madrid.data.dataSource.remote.response.series.AiringTodayTvShowsResponse
-import com.madrid.data.dataSource.remote.response.series.OnAirTvShowsResponse
-import com.madrid.data.dataSource.remote.response.series.OnAirTvShowsResult
-import com.madrid.data.dataSource.remote.response.series.RecommendedSeriesResponse
-import com.madrid.data.dataSource.remote.response.series.RecommendedSeriesResult
-import com.madrid.data.dataSource.remote.response.series.TopRatedSeriesResponse
-import com.madrid.data.dataSource.remote.response.series.TopRatedSeriesResults
+import com.madrid.data.dataSource.remote.dto.series.AiringTodaySeriesResult
+import com.madrid.data.dataSource.remote.dto.series.AiringTodayTvShowsResponse
+import com.madrid.data.dataSource.remote.dto.series.OnAirTvShowsResponse
+import com.madrid.data.dataSource.remote.dto.series.OnAirTvShowsResult
+import com.madrid.data.dataSource.remote.dto.series.RecommendedSeriesResponse
+import com.madrid.data.dataSource.remote.dto.series.RecommendedSeriesResult
+import com.madrid.data.dataSource.remote.dto.series.TopRatedSeriesResponse
+import com.madrid.data.dataSource.remote.dto.series.TopRatedSeriesResults
 import com.madrid.domain.entity.Artist
 import com.madrid.domain.entity.Episode
 import com.madrid.domain.entity.Genre
